+++ enjoy/src/renderer/components/misc/index.ts
@@ -8,5 +8,6 @@ export * from "./github-login-form";
 export * from "./mixin-login-form";
 export * from "./no-records-found";
 export * from "./page-placeholder";
+export * from "./player";
 export * from "./sidebar";
 export * from "./wavesurfer-player";
